provide support for options with addModel and bug fix trace style
It wasn't possible to provide an options object to addModel.
This is now available and passed on to the parser.
I changed the default behavior of mol2 to keep hydrogens.

Michal Jamroz reported a bug with trace cartoons when the file consists
of only CAs.  Have fixed the bug and added a thickness option