More animation support; bf
Labels and shapes can now be specified with a frame id in
which case they will only be displayed when animating that frame.

vibrate now has an option to take an arrow spec to draw arrows
that show the difference vector (they move).  I don't like this,
but there is no accounting for taste.

addResLabels now takes an option to apply by frame so labels move
during animations.

Fixed a bug buried in the WebGL code with removing objects.