Better antialiasing
Well, not really.  I realized that devicePixelRatio changes with the
zoom level and by rendering to twice the pixels we can get a less
pixelated result.  Also, antialiasing is set on the canvas.
Unfortunately, these realizations came about because it turns out that I
created all the test reference images with my browser slightly zoomed
in.  Since they need to be regenerated anyway, I'm going to render to a
higher resolution as long as it isn't too slow on my laptop/phone.