Normalization of symmetries and fancier unit cell.
Apparently crystalographers don't bother specify symmetry
transformations that put molecules in nice places - you are suppose to
translate the center of the molecule into the unit cell to get a good
representation.  This is now possible with the normalizeAssembly option.

I've also made addUnitCell much fancier with fully customizable styling
of the box, axes arrows, and labels.