antialias fix
Doubling resolution causes iOS Safari to crash, but only when antialias
setting is set on canvas.  Since I can't visual detect what this setting
is doing, disable it.  Previously I had hackily disabled resolution
doubling when a retina caliber display was detected.  This apparently
results in jaggies on linux (which doesn't make a lot of sense to me -
the display should be high resolution, but oh well).