Add ability to set start frame when animating
Note: this is rough and untested for backwards and back-and-forth. It also may cause problems when trying to start on a frame that's too high, and will likely crash if this is attempted.

Adds new "startFrame" argument to viewer.animate function, which specifies what frame to start the animation on.