Robust glcheck
Fix blacklisting problems and take screenshot of only canvas, not whole
page.  This allows us to use the glcheck generated images as the
reference images for auto/tests as well as glcheck which is preferred
since glcheck will be what is run in CI.

This requires using a fork of glcheck to support canvas snapshots and
view resizing.

Avoid copying >100MB of data to run glcheck.

Fix a few issues that popped up in getting this all to work.