Regenerate lost contexts
Issue #795.  When a context is lost, go through all visible contexts
(up to the first 16) and regenerate them if needed.

If a browser supports fewer than 16 contexts and too many contexts are
on the screen there will be an infinite loop of context regeneration,
but at least on the desktop 16 seems like a safe lower bound.