Better selection.
Issue #804.  Model selection was only being processed at the top level
when the selection was dispatched to the models in the viewer.  Now
process in the model as well to properly handle model selection in
nested statements like or/and.

This requires giving models access to their viewer and so I've put in
some minor api breakage where model methods no longer need to be passed
a viewer.